requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiocontextvars==0.2.2; python_version < '3.7'
asgiref==3.4.1; python_version >= '3.6'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
autobahn==21.2.1; python_version >= '3.6'
automat==20.2.0
cffi==1.14.6
channels==3.0.4
constantly==15.1.0
contextvars==2.4; python_version < '3.7'
cryptography==3.4.7; python_version >= '3.6'
daphne==3.0.2; python_version >= '3.6'
django-rest-framework==0.1.0
django==3.2.6
djangorestframework==3.12.4; python_version >= '3.5'
hyperlink==21.0.0
idna==3.2
immutables==0.16; python_version >= '3.6'
incremental==21.3.0
loguru==0.5.3
pillow==8.3.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyopenssl==20.0.1
pytz==2021.1
service-identity==21.1.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlparse==0.4.1; python_version >= '3.5'
twisted[tls]==21.7.0; python_full_version >= '3.6.7'
txaio==21.2.1; python_version >= '3.6'
typing-extensions==3.10.0.0; python_version < '3.8' and python_version < '3.8'
zope.interface==5.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
